docs: Update gitignore to ignore Kotlin errors directory
This commit updates the `.gitignore` file to specifically ignore the `/.kotlin/errors` directory.

Previously, the `.gitignore` file contained the entry `.kotlin`, which is incorrect syntax according to the gitignore documentation [1]. This change replaces that entry with the correct path `/.kotlin/errors` to ensure that the Kotlin error log directory is properly ignored.